Update patch set 3
Patch Set 3: Do not submit

Sorry, only the maintainers can submit prebuilts.

In any case, this is unnecessary. The audio distortion is a common misconfiguration in bcm4325 (I have this in a couple of other devices) and it's easily fixed with hcitool; I'll push up a patch in a few minutes.

Patch-set: 3