Update patch set 2
Patch Set 2:

The problem is thus:

SND_DEVICE_HEADSET has mono audio and active handset mic.
SND_DEVICE_HEADSET_STEREO has stereo audio but MUTED handset mic.

For phone calls we would want SND_DEVICE_HEADSET so that we can answer a phone call when using headphones. When listening to media we would want SND_DEVICE_HEADSET_STEREO so we can listen to stereo media.

Unfortunately in CM7 playback of media uses the same routing as for in a phone call. We cannot have stereo output when listening to media and have the ability to answer a call using headphones and also use the handset mic.

VOIP calls (Skype etc.) in CM7 do use a different routing so we can use headphones and the handset mic there.

CM9 apparently handles routing differently as stereo media output plus active handset mic during phone calls works correctly. The problem isn't in the device libaudio as the routing is identical in CM7 and CM9.

Now it's time to look at CM9 and see what the difference is. I suspect it's connected to the reason why headphones are detected as a wired headset in CM7.

Patch-set: 2